// src/types/filters.ts
export enum OrderStatusFilter {
  PENDING = "PENDING",
  ACCEPTED = "ACCEPTED",
  READY_FOR_PICKUP = "READY_FOR_PICKUP",
  COMPLETED = "COMPLETED",
  REJECTED = "REJECTED",
  CANCELLED = "CANCELLED",
}

export enum TimeFilterOption {
  LAST_30_MIN = "LAST_30_MIN",
  LAST_1_HOUR = "LAST_1_HOUR",
  LAST_3_HOUR = "LAST_3_HOUR",
  TODAY = "TODAY",
  LAST_WEEK = "LAST_WEEK",
  LAST_1_MONTH = "LAST_1_MONTH",
  THIS_MONTH = "THIS_MONTH",
  THIS_YEAR = "THIS_YEAR",
  CUSTOM = "CUSTOM",
}